Review by AJCM

+++ Content.Server/Xenoarchaeology/Artifact/XenoArtifactSystem.ProcGen.cs
@@ -156,7 +156,7 @@ public sealed partial class XenoArtifactSystem
         // Default to one node if we had shenanigans and ended up with weird layer counts.
         var nodeCount = 1;
         if (layerMax >= layerMin)
-            nodeCount = RobustRandom.Next(layerMin, layerMax + 1); // account for non-inclusive max
+            nodeCount = RobustRandom.Next((int) layerMin, (int) layerMax + 1); // account for non-inclusive max
 
         segmentSize -= nodeCount;
         var nodes = new List<Entity<XenoArtifactNodeComponent>>();
@@ -204,7 +204,7 @@ public sealed partial class XenoArtifactSystem
         var segmentMin = ent.Comp.SegmentSize.Min;
         var segmentMax = Math.Min(ent.Comp.SegmentSize.Max, Math.Max(nodeCount / 2, segmentMin));
 
-        var segmentSize = RobustRandom.Next(segmentMin, segmentMax + 1); // account for non-inclusive max
+        var segmentSize = RobustRandom.Next((int) segmentMin, (int) segmentMax + 1); // account for non-inclusive max
         var remainder = nodeCount - segmentSize;
 
         // If our next segment is going to be undersized, then we just absorb it into this segment.
